优化敌人波次生成与关卡配置逻辑，修复敌人移动模式、子弹类型枚举及对象池复用相关问题

+++ Assets/Scripts/Game/Enemy/EnemySpawnConfig.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class EnemySpawnConfig
 {
     [Header("生成时间")]
-    public int spawnTime;// 基于音乐时间
+    public float spawnTime;// 基于音乐时间
     
     [Header("生成数量")]
     public int spawnCount = 1;// 生成敌人的数量
